% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.gb.R
\name{extract.gb}
\alias{extract.gb}
\title{Extracts and returns a specific item from .gb records}
\usage{
extract.gb(x, Item)
}
\arguments{
\item{x}{The name of the list containing the records.}

\item{Item}{character. The part to extract. It have to match case.}
}
\value{
Returns a list containing the data
}
\description{
This function regroup every items found in a list if they correspond to the character string specified. Works for general parts (ACCESSION, SOURCE...) and parts of FEATURES (source, gene, CDS...)
}
\examples{
\dontrun{
extract.gb(x, "gene")
}

}
